refactor(imports): update import paths for wallet balance files
