+++ lib/presentation/home_page/dapps/wallet_balance/wallet_balance_page.dart
