+++ lib/presentation/home_page/home_ui.dart
@@ -3,7 +3,7 @@ import "package:flutter/material.dart";
 import "package:flutter/rendering.dart";
 
 import "package:hnotes/presentation/drawer/drawer_ui.dart";
-import "package:hnotes/presentation/home_page/dapps/wallet_balance_card.dart";
+import "package:hnotes/presentation/home_page/dapps/wallet_balance/wallet_balance_card.dart";
 import "package:hnotes/presentation/home_page/header/header_widget.dart";
 import "package:hnotes/presentation/home_page/items/build_item_list.dart";
 import "package:hnotes/presentation/home_page/header/drawer_icon_widget.dart";
